Add a command-line system to ParserGenApp.
This extends ParserGenApp so that it can be run with command line arguments,
in which case it does not show the configuration window, but it just runs the
parser generator instead.

This moves some of the generation code out of PGDocument and into a new
PGGenerator.  (It gets ARCified and tidied up at the same time.)