Make it possible for a PKParser subclass to use ARC.
To do this, all the references to self inside blocks need to be
weak, otherwise nothing gets freed.

This has been implemented with a couple of macros that take different
values depending on whether the code is compiled with ARC.
All references to self that would appear inside a block (i.e.
just about any of the templated fragments) need to use PKParser_weakSelf
instead of self, and each of the method templates needs to include
PKParser_weakSelfDecl.

Note that the subclass needs to be compiled with -fobjc-arc-exceptions
to avoid leaks, because PKParser uses exceptions for control flow heavily.