Fix PKReader.offset in the case where streams are being used.
The unread functionality added in 9b7987469e used self.offset as
the offset into self.buffer.  That's no good though, because
self.offset is externally expected to be the offset into the
whole input (i.e. the stream) and not the offset into an internal
buffer.

Fix this by adding a separate self.bufOffset.